pyproject.toml
[project]
name = "mcp-web-finder"
version = "0.1.0"
description = "Un servicio MCP para búsqueda y análisis web avanzado"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.6.1",
    "crawl4ai>=0.6.3",
    "langchain-google-community>=2.0.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "colorama>=0.4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
] 